backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
